requirements.txt
Django==1.6
Pillow==2.5.1
distribute==0.6.31
django-threadedcomments
django-bootstrap3==4.8.2
django-registration==1.0
sorl-thumbnail==11.12.1b
python-social-auth
wsgiref==0.1.2
uuid==1.30
django-watson==1.1.3